solve all problem

+++ exercises/23_conversions/using_as.rs
@@ -1,11 +1,14 @@
 // Type casting in Rust is done via the usage of the `as` operator.
 // Note that the `as` operator is not only used when type casting. It also helps
 // with renaming imports.
+// Rust中的类型转换是通过使用 'as' 运算符完成的。
+// 请注意，“as” 运算符不仅在类型转换时使用。它也有助于
+// 通过重命名导入。
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
     // TODO: Make a conversion before dividing.
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
